+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -20124,7 +20124,7 @@ $ES_Completers = {
             ($_ -eq "Update-ESDomainConfig/EBSOptions_VolumeType")
         }
         {
-            $v = "gp2","io1","standard"
+            $v = "gp2","gp3","io1","standard"
             break
         }
 
@@ -36460,7 +36460,7 @@ $OS_Completers = {
             ($_ -eq "Update-OSDomainConfig/EBSOptions_VolumeType")
         }
         {
-            $v = "gp2","io1","standard"
+            $v = "gp2","gp3","io1","standard"
             break
         }