+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Get-EC2SecurityGroup-Cmdlet.cs
@@ -37,7 +37,11 @@ namespace Amazon.PowerShell.Cmdlets.EC2
     /// EC2 security groups</a> in the <i>Amazon Elastic Compute Cloud User Guide</i> and
     /// <a href="https://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_SecurityGroups.html">Security
     /// groups for your VPC</a> in the <i>Amazon Virtual Private Cloud User Guide</i>.
-    /// </para><br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
+    /// </para><note><para>
+    /// We are retiring EC2-Classic on August 15, 2022. We recommend that you migrate from
+    /// EC2-Classic to a VPC. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/vpc-migrate.html">Migrate
+    /// from EC2-Classic to a VPC</a> in the <i>Amazon Elastic Compute Cloud User Guide</i>.
+    /// </para></note><br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
     /// </summary>
     [Cmdlet("Get", "EC2SecurityGroup")]
     [OutputType("Amazon.EC2.Model.SecurityGroup")]