+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/New-EC2InstanceExportTask-Cmdlet.cs
@@ -102,7 +102,10 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         /// <summary>
         /// <para>
         /// <para>The Amazon S3 bucket for the destination image. The destination bucket must exist
-        /// and grant WRITE and READ_ACP permissions to the Amazon Web Services account <code>vm-import-export@amazon.com</code>.</para>
+        /// and have an access control list (ACL) attached that specifies the Region-specific
+        /// canonical account ID for the <code>Grantee</code>. For more information about the
+        /// ACL to your S3 bucket, see <a href="https://docs.aws.amazon.com/vm-import/latest/userguide/vmexport.html#vmexport-prerequisites">Prerequisites</a>
+        /// in the VM Import/Export User Guide.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]