+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Revoke-EC2SecurityGroupIngress-Cmdlet.cs
@@ -49,7 +49,11 @@ namespace Amazon.PowerShell.Cmdlets.EC2
     /// </para><para>
     /// Rule changes are propagated to instances within the security group as quickly as possible.
     /// However, a small delay might occur.
-    /// </para>
+    /// </para><note><para>
+    /// We are retiring EC2-Classic on August 15, 2022. We recommend that you migrate from
+    /// EC2-Classic to a VPC. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/vpc-migrate.html">Migrate
+    /// from EC2-Classic to a VPC</a> in the <i>Amazon Elastic Compute Cloud User Guide</i>.
+    /// </para></note>
     /// </summary>
     [Cmdlet("Revoke", "EC2SecurityGroupIngress", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("Amazon.EC2.Model.RevokeSecurityGroupIngressResponse")]