+++ modules/AWSPowerShell/Cmdlets/Elasticsearch/AWS.Tools.Elasticsearch.Completers.psm1
@@ -150,7 +150,7 @@ $ES_Completers = {
             ($_ -eq "Update-ESDomainConfig/EBSOptions_VolumeType")
         }
         {
-            $v = "gp2","io1","standard"
+            $v = "gp2","gp3","io1","standard"
             break
         }