+++ modules/AWSPowerShell/Cmdlets/Elasticsearch/Basic/Update-ESDomainConfig-Cmdlet.cs
@@ -342,7 +342,7 @@ namespace Amazon.PowerShell.Cmdlets.ES
         #region Parameter EBSOptions_Iops
         /// <summary>
         /// <para>
-        /// <para>Specifies the IOPD for a Provisioned IOPS EBS volume (SSD).</para>
+        /// <para>Specifies the IOPS for Provisioned IOPS And GP3 EBS volume (SSD).</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -520,6 +520,16 @@ namespace Amazon.PowerShell.Cmdlets.ES
         public System.String[] VPCOptions_SubnetId { get; set; }
         #endregion
         
+        #region Parameter EBSOptions_Throughput
+        /// <summary>
+        /// <para>
+        /// <para>Specifies the Throughput for GP3 EBS volume (SSD).</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.Int32? EBSOptions_Throughput { get; set; }
+        #endregion
+        
         #region Parameter DomainEndpointOptions_TLSSecurityPolicy
         /// <summary>
         /// <para>
@@ -695,6 +705,7 @@ namespace Amazon.PowerShell.Cmdlets.ES
             context.DryRun = this.DryRun;
             context.EBSOptions_EBSEnabled = this.EBSOptions_EBSEnabled;
             context.EBSOptions_Iops = this.EBSOptions_Iops;
+            context.EBSOptions_Throughput = this.EBSOptions_Throughput;
             context.EBSOptions_VolumeSize = this.EBSOptions_VolumeSize;
             context.EBSOptions_VolumeType = this.EBSOptions_VolumeType;
             context.ColdStorageOptions_Enabled = this.ColdStorageOptions_Enabled;
@@ -1124,6 +1135,16 @@ namespace Amazon.PowerShell.Cmdlets.ES
                 request.EBSOptions.Iops = requestEBSOptions_eBSOptions_Iops.Value;
                 requestEBSOptionsIsNull = false;
             }
+            System.Int32? requestEBSOptions_eBSOptions_Throughput = null;
+            if (cmdletContext.EBSOptions_Throughput != null)
+            {
+                requestEBSOptions_eBSOptions_Throughput = cmdletContext.EBSOptions_Throughput.Value;
+            }
+            if (requestEBSOptions_eBSOptions_Throughput != null)
+            {
+                request.EBSOptions.Throughput = requestEBSOptions_eBSOptions_Throughput.Value;
+                requestEBSOptionsIsNull = false;
+            }
             System.Int32? requestEBSOptions_eBSOptions_VolumeSize = null;
             if (cmdletContext.EBSOptions_VolumeSize != null)
             {
@@ -1466,6 +1487,7 @@ namespace Amazon.PowerShell.Cmdlets.ES
             public System.Boolean? DryRun { get; set; }
             public System.Boolean? EBSOptions_EBSEnabled { get; set; }
             public System.Int32? EBSOptions_Iops { get; set; }
+            public System.Int32? EBSOptions_Throughput { get; set; }
             public System.Int32? EBSOptions_VolumeSize { get; set; }
             public Amazon.Elasticsearch.VolumeType EBSOptions_VolumeType { get; set; }
             public System.Boolean? ColdStorageOptions_Enabled { get; set; }