+++ modules/AWSPowerShell/Cmdlets/LookoutforVision/Basic/Find-LFVAnomaly-Cmdlet.cs
@@ -34,6 +34,8 @@ namespace Amazon.PowerShell.Cmdlets.LFV
     /// <para>
     /// The response from <code>DetectAnomalies</code> includes a boolean prediction that
     /// the image contains one or more anomalies and a confidence value for the prediction.
+    /// If the model is an image segmentation model, the response also includes segmentation
+    /// information for each type of anomaly found in the image.
     /// </para><note><para>
     /// Before calling <code>DetectAnomalies</code>, you must first start your model with
     /// the <a>StartModel</a> operation. You are charged for the amount of time, in minutes,
@@ -41,6 +43,9 @@ namespace Amazon.PowerShell.Cmdlets.LFV
     /// If you are not using a model, use the <a>StopModel</a> operation to stop your model.
     /// 
     /// </para></note><para>
+    /// For more information, see <i>Detecting anomalies in an image</i> in the Amazon Lookout
+    /// for Vision developer guide.
+    /// </para><para>
     /// This operation requires permissions to perform the <code>lookoutvision:DetectAnomalies</code>
     /// operation.
     /// </para>