+++ modules/AWSPowerShell/Cmdlets/LookoutforVision/Basic/Start-LFVModelPackagingJob-Cmdlet.cs
@@ -63,9 +63,12 @@ namespace Amazon.PowerShell.Cmdlets.LFV
         #region Parameter TargetPlatform_Accelerator
         /// <summary>
         /// <para>
-        /// <para>The target accelerator for the model. NVIDIA (Nvidia graphics processing unit) is
-        /// the only accelerator that is currently supported. You must also specify the <code>gpu-code</code>,
-        /// <code>trt-ver</code>, and <code>cuda-ver</code> compiler options. </para>
+        /// <para>The target accelerator for the model. Currently, Amazon Lookout for Vision only supports
+        /// NVIDIA (Nvidia graphics processing unit) and CPU accelerators. If you specify NVIDIA
+        /// as an accelerator, you must also specify the <code>gpu-code</code>, <code>trt-ver</code>,
+        /// and <code>cuda-ver</code> compiler options. If you don't specify an accelerator, Lookout
+        /// for Vision uses the CPU for compilation and we highly recommend that you use the <a>GreengrassConfiguration$CompilerOptions</a>
+        /// field. For example, you can use the following compiler options for CPU: </para><ul><li><para><code>mcpu</code>: CPU micro-architecture. For example, <code>{'mcpu': 'skylake-avx512'}</code></para></li><li><para><code>mattr</code>: CPU flags. For example, <code>{'mattr': ['+neon', '+vfpv4']}</code></para></li></ul>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -112,8 +115,7 @@ namespace Amazon.PowerShell.Cmdlets.LFV
         /// <summary>
         /// <para>
         /// <para>Additional compiler options for the Greengrass component. Currently, only NVIDIA Graphics
-        /// Processing Units (GPU) are supported. If you specify <code>TargetPlatform</code>,
-        /// you must specify <code>CompilerOptions</code>. If you specify <code>TargetDevice</code>,
+        /// Processing Units (GPU) and CPU accelerators are supported. If you specify <code>TargetDevice</code>,
         /// don't specify <code>CompilerOptions</code>.</para><para>For more information, see <i>Compiler options</i> in the Amazon Lookout for Vision
         /// Developer Guide. </para>
         /// </para>