+++ modules/AWSPowerShell/Cmdlets/OpenSearchService/Basic/New-OSDomain-Cmdlet.cs
@@ -343,7 +343,7 @@ namespace Amazon.PowerShell.Cmdlets.OS
         #region Parameter EBSOptions_Iops
         /// <summary>
         /// <para>
-        /// <para>The IOPD for a Provisioned IOPS EBS volume (SSD).</para>
+        /// <para>The IOPS for Provisioned IOPS And GP3 EBS volume (SSD).</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -527,6 +527,16 @@ namespace Amazon.PowerShell.Cmdlets.OS
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
         public Amazon.OpenSearchService.Model.Tag[] TagList { get; set; }
+        #endregion
+        
+        #region Parameter EBSOptions_Throughput
+        /// <summary>
+        /// <para>
+        /// <para>The Throughput for GP3 EBS volume (SSD).</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.Int32? EBSOptions_Throughput { get; set; }
         #endregion
         
         #region Parameter DomainEndpointOptions_TLSSecurityPolicy
@@ -714,6 +724,7 @@ namespace Amazon.PowerShell.Cmdlets.OS
             #endif
             context.EBSOptions_EBSEnabled = this.EBSOptions_EBSEnabled;
             context.EBSOptions_Iops = this.EBSOptions_Iops;
+            context.EBSOptions_Throughput = this.EBSOptions_Throughput;
             context.EBSOptions_VolumeSize = this.EBSOptions_VolumeSize;
             context.EBSOptions_VolumeType = this.EBSOptions_VolumeType;
             context.EncryptionAtRestOptions_Enabled = this.EncryptionAtRestOptions_Enabled;
@@ -1273,6 +1284,16 @@ namespace Amazon.PowerShell.Cmdlets.OS
                 request.EBSOptions.Iops = requestEBSOptions_eBSOptions_Iops.Value;
                 requestEBSOptionsIsNull = false;
             }
+            System.Int32? requestEBSOptions_eBSOptions_Throughput = null;
+            if (cmdletContext.EBSOptions_Throughput != null)
+            {
+                requestEBSOptions_eBSOptions_Throughput = cmdletContext.EBSOptions_Throughput.Value;
+            }
+            if (requestEBSOptions_eBSOptions_Throughput != null)
+            {
+                request.EBSOptions.Throughput = requestEBSOptions_eBSOptions_Throughput.Value;
+                requestEBSOptionsIsNull = false;
+            }
             System.Int32? requestEBSOptions_eBSOptions_VolumeSize = null;
             if (cmdletContext.EBSOptions_VolumeSize != null)
             {
@@ -1508,6 +1529,7 @@ namespace Amazon.PowerShell.Cmdlets.OS
             public System.String DomainName { get; set; }
             public System.Boolean? EBSOptions_EBSEnabled { get; set; }
             public System.Int32? EBSOptions_Iops { get; set; }
+            public System.Int32? EBSOptions_Throughput { get; set; }
             public System.Int32? EBSOptions_VolumeSize { get; set; }
             public Amazon.OpenSearchService.VolumeType EBSOptions_VolumeType { get; set; }
             public System.Boolean? EncryptionAtRestOptions_Enabled { get; set; }